requirements.txt
fitz
pymupdf
pdfplumber
python-docx
pandas
PIL